+++ src/main/java/Cortex.java
@@ -1,19 +1,19 @@
 import java.util.ArrayList;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 public class Cortex {
     public ArrayList<Task> list;
     private Storage storage;
     private Ui ui;
+    private Parser parser;
 
 
     public Cortex() {
         this.storage = new Storage("./data/duke.txt");
         this.list = storage.loadtasks();
         ui = new Ui();
+        parser = new Parser();
 
     }
     public static void main(String[] args) {
@@ -26,9 +26,7 @@ public class Cortex {
      */
     public void run() {
 
-
         ui.printHello();
-
         String command = ui.readCommand();
 
 
@@ -53,62 +51,21 @@ public class Cortex {
                 int c = 0;
                 if (command.startsWith("todo")) {
                     c++;
-                    try {
-                        String s = command.substring(5).trim();
-                        task = new Todo(s);
-                        isValidTask = true;
-                    } catch (Exception e){
-                        ui.printError("INVALID TODO TASK");
-                    }
+                    task = parser.parseTodoCommand(command);
 
                 } else if (command.startsWith("deadline")) {
                     c++;
-                    try {
-                        String s = command.substring(9).trim();
-                        String[] bySplit = s.split("/by");
-                        String desc = bySplit[0].trim();
-                        String by = bySplit[1].trim();
-
-                        try {
-                            LocalDateTime byDateTime = parseDateTime(by);
-                            task = new Deadline(desc, byDateTime);
-                        } catch (DateTimeParseException e) {
-                            task = new Deadline(desc, by);
-                        }
-
-                        isValidTask = true;
-                    } catch (Exception e) {
-                        ui.printError("INVALID DEADLINE TASK");
-                    }
+                    task = parser.parseDeadlineCommand(command);
 
                 } else if (command.startsWith("event")) {
                     c++;
-                    try {
-                        String s = command.substring(6).trim();
-                        String[] fromSplit = s.split("/from");
-                        String desc = fromSplit[0].trim();
-
-                        String[] toSplit = fromSplit[1].split("/to");
-                        String from = toSplit[0];
-                        String to = toSplit[1];
-
-                        try {
-                            LocalDateTime fromDateTime = parseDateTime(from);
-                            LocalDateTime toDateTime = parseDateTime(to);
-                            task = new Event(desc, fromDateTime, toDateTime);
-                        } catch (DateTimeParseException e) {
-                            task = new Event(desc, from, to);
-                        }
-                        
-                        isValidTask = true;
-                    } catch (Exception e) {
-                        ui.printError("INVALID EVENT TASK");
-                    }
+                    task = parser.parseEventCommand(command);
                 }
 
-                if(isValidTask) {
+                if(task != null) {
                     addTask(task);
-                } else if (c == 0){
+                }
+                if (c == 0){
                     ui.printError("INVALID TASK TYPE");
                 }
 
@@ -117,11 +74,8 @@ public class Cortex {
             ui.printHorizontalLine();
             command = ui.readCommand();
         }
-
         ui.printBye();
     }
-
-
 
 
     /**
@@ -131,12 +85,12 @@ public class Cortex {
      */
     public void markTask(String command){
         try {
-            int  i= Integer.parseInt(command.substring(5).trim()) - 1;
+            int  i = parser.parseMarkCommand(command);
+
             if(i < 0 || i >= list.size()) {
                 ui.printError("Invalid Task! Cannot mark task.");
             } else {
                 Task t = list.get(i);
-
                 t.markAsDone();
                 ui.printMarked(t);
             }
@@ -156,13 +110,12 @@ public class Cortex {
      */
     public void unmarkTask(String command) {
         try {
-            int i = Integer.parseInt(command.substring(7).trim()) - 1;
+            int i = parser.parseUnmarkCommand(command);
 
             if(i < 0 || i >= list.size()) {
                 ui.printError("Invalid Task! Cannot unmark task.");
             } else {
                 Task t = list.get(i);
-
                 t.unmarkAsDone();
                 ui.printUnmarked(t);
             }
@@ -183,14 +136,14 @@ public class Cortex {
      */
     public void deleteTask(String command) {
         try {
-            int i = Integer.parseInt(command.substring(7).trim()) - 1;
+            int i = parser.parseDeleteCommand(command);
 
             if(i < 0 || i >= list.size()) {
                 ui.printError("Invalid Task! Cannot delete task.");
             } else {
                 Task t = list.get(i);
                 list.remove(i);
-               ui.printDeletedTask(t, list);
+                ui.printDeletedTask(t, list);
             }
 
         } catch (NumberFormatException e) {
@@ -213,33 +166,4 @@ public class Cortex {
         storage.saveTasks(list);
     }
 
-    /**
-     * Returns a string as LocalDateTime object.
-     *
-     * @param timeString to store the time as a string.
-     * @throws DateTimeParseException if the timeString is invalid time format.
-     */
-    public LocalDateTime parseDateTime(String timeString) throws DateTimeParseException {
-        timeString = timeString.trim();
-
-        if (timeString.contains(" ")) {
-            String[] parts = timeString.split(" ");
-            String dd = parts[0];
-            String tt = parts[1];
-
-            DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("d/M/yyyy");
-            LocalDate date = LocalDate.parse(dd, dateFormatter);
-
-            if (parts[1].length() == 4) {
-                int hour = Integer.parseInt(parts[1].substring(0, 2));
-                int minute = Integer.parseInt(parts[1].substring(2));
-                return LocalDateTime.of(date, java.time.LocalTime.of(hour, minute));
-            } else {
-                throw new DateTimeParseException("Invalid time!", timeString, 0);
-            }
-        } else {
-            DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("d/M/yyyy");
-            return LocalDate.parse(timeString, dateFormatter).atStartOfDay();
-        }
-    }
 }
